Lower threshold to 0.15 from 0.25
The whole point of the running total is make more recent folks and data relevant,
and this threshold change is a big part of it.  We'll have some issues in the
future as the lifetime total of games increases significantly (still have
to have played in six games to show up there, which is a lot) - maybe that
changes to a specific value rather than a percentage?